Move function generator to own module
